requirements.txt
streamlit==1.56.0
google-genai==1.70.0
pandas==3.0.2
plotly==6.6.0
python-dotenv==1.2.2
supabase==2.28.3
